Move tool invocation parts until toolInvocationParts/
